refactor(ui): move model/specialist and skills/plugins settings domains out of App
App keeps only construction and wiring: ModelSettingsState owns the model
and specialist form signals plus save/validate/test handlers, and
ExtensionsState owns the skills/plugins pane state plus install/enable/
remove handlers. Handlers in main.rs become one-line delegations, and the
cross-domain signals (models, settings, busy flag, install modal, locale)
are injected at construction.
