+++ ui/src/app_support/mod.rs
@@ -31,11 +31,13 @@ mod chat_stream;
 mod composer;
 mod context_usage;
 mod dnd;
+mod extensions;
 mod files;
 mod issue_report;
 mod markdown;
 mod messages;
 mod modals;
+mod model_settings;
 mod palettes;
 mod pane_layout;
 mod prefs;
@@ -54,11 +56,13 @@ pub(crate) use chat_stream::*;
 pub(crate) use composer::*;
 pub(crate) use context_usage::*;
 pub(crate) use dnd::*;
+pub(crate) use extensions::*;
 pub(crate) use files::*;
 pub(crate) use issue_report::*;
 pub(crate) use markdown::*;
 pub(crate) use messages::*;
 pub(crate) use modals::*;
+pub(crate) use model_settings::*;
 pub(crate) use palettes::*;
 pub(crate) use pane_layout::*;
 pub(crate) use prefs::*;
